Move events to storage for better cohesion.
