+++ src/main/java/seedu/multitasky/ui/StatusBarFooter.java
@@ -12,7 +12,7 @@ import javafx.fxml.FXML;
 import javafx.scene.layout.Region;
 import seedu.multitasky.commons.core.LogsCenter;
 import seedu.multitasky.commons.events.model.EntryBookChangedEvent;
-import seedu.multitasky.commons.events.model.EntryBookToUndoEvent;
+import seedu.multitasky.commons.events.storage.EntryBookToUndoEvent;
 
 /**
  * A ui for the status bar that is displayed at the footer of the application.
